kernel_riscv

Module trap

Structs

TrapFrame

Functions

early_trap_handler 🔒

install_early_handler

Install the early trap handler - this should be installed very early in the kernel, and swapped out once the task-scheduling infrastructure is up, as it uses a more advanced mechanism to handle traps from both S-mode and U-mode.

install_full_handler

Install the proper trap handler. This handler is able to take traps from both S-mode and U-mode, but requires the sscratch context to be correctly installed to facilitate switching to the kernel’s stack correctly. It therefore cannot be used during early initialization.

trap_handler 🔒

trap_handler_shim 🔒